Class In_buf.bufferized

inherit Iostream__.In.t

method fill_buf : unit -> Iostream__.Slice.t

returns a slice into the bic's internal buffer, and ensures it's empty only if bic.ic is empty.

method consume : int -> unit

Consume n bytes from the inner buffer. This is only valid if the last call to fill_buf returned a slice with at least n bytes.

method input : bytes -> int -> int -> int

Default implementation of input using fill_buf