Class In_buf.of_in_channel

inherit Iostream__.In.t

method fill_buf : unit -> Iostream__.Slice.t

ic#fill_buf() returns a slice into the ic's internal buffer, and ensures it's empty only if ic.ic is empty. In other words, the invariant is that this only returns an empty slice if the input stream is exhausted.

method consume : int -> unit

Consume n bytes from the inner buffer. This is only valid if the last call to fill_buf returned a slice with at least n bytes.