Class In_buf.t_from_refill

An implementation of t that relies on a Slice.t as buffer, along with a refill method to fill the buffer with data when it's entirely consumed.

method virtual refill : Iostream__.Slice.t -> unit

Implementation of the stream: this takes a slice, resets its offset, and fills it with bytes. It must write at least one byte in the slice, unless the underlying input has reached its end.

inherit t