Class Out_buf.t_from_output

Make a bufferized output from a non bufferized output+close.

parameter bytes
the buffer to use. It's owned by this channel as long as the channel exists.

inherit t

method private virtual output_underlying : bytes -> int -> int -> unit

Emit these private bytes, unbufferized

method private virtual close_underlying : unit -> unit

Close the underlying output. The bufferized output will flush and then call this.