Class type Out_buf.t

An output stream, ie. a place into which we can write bytes, with a buffer to amortize the cost of operations.

This can be a Buffer.t, an out_channel, a Unix.file_descr, etc.

inherit Iostream__.Out.t

method output_char : char -> unit

Output a single char

method flush : unit -> unit

Flush underlying buffer